Placed into the right place
