+++ src/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Library Admin Interface</title>
+    <title>Library Admin</title>
     <link rel="stylesheet" href="style.css">
 </head>
 <body>
